fix: do not bail fast when fetching an endpoint's cancellation token
When receiving the last data frame the remote endpoint is closed causing it's cancellation source to cancel. The same cancellation token is used to bail when waiting for more data. However, data will be fetched synchronously if data is available, which it won't be able to do if the cancellation token throws when it is fetched.